Improving app moduling
